+++ things/clients/cache/doc.go
@@ -1,0 +1,6 @@
+// Copyright (c) Mainflux
+// SPDX-License-Identifier: Apache-2.0
+
+// Package cache contains the domain concept definitions needed to
+// support Mainflux things cache service functionality.
+package cache
