+++ things/clients/cache/things.go
@@ -1,7 +1,7 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-package events
+package cache
 
 import (
 	"context"
